+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Update the readme file list of known issues
 * Delete any existing `.vscode_*_done` temporary files created by the `LogtalkTerminal` class at initialization time
+* Faster check for minimum required Logtalk version at extension activation
 
 ## [0.64.0]